+++ app/javascript/controllers/todos_controller.js
@@ -1,0 +1,8 @@
+import ApplicationController from './application_controller';
+
+export default class extends ApplicationController {
+  cancelEdit(event) {
+    if (event.type === 'keydown' && !['Escape', 'Esc'].includes(event.key)) return;
+    this.stimulate('TodosReflex#cancel_edit');
+  }
+}
