+++ app/views/todos/index.html.erb
@@ -1,17 +1,17 @@
-<section class="todoapp">
+<section class="todoapp" data-controller="visual-effects">
   <header class="header">
     <h1><small><%= truncate "#{session[:filter]} todos", length: 12 %></small></h1>
     <input <%= "autofocus " unless @edit_id %> class="new-todo" placeholder="What needs to be done?"
-      data-controller="todo" data-reflex="change->TodosReflex#create" data-room="<%= session.id %>">
+      data-reflex="change->TodosReflex#create" data-room="<%= session.id %>">
   </header>
 
   <% if @all_todos.any? %>
     <section class="main">
       <input class="toggle-all" type="checkbox">
-      <label data-controller="todo" data-reflex="click->TodosReflex#toggle_all" data-room="<%= session.id %>">
+      <label data-reflex="click->TodosReflex#toggle_all" data-room="<%= session.id %>">
         Mark all as complete
       </label>
-      <ul class="todo-list">
+      <ul class="todo-list" data-target="visual-effects.list">
         <%= render partial: "todo", collection: @filtered_todos %>
       </ul>
     </section>
